Fix client UUID gen, and some warnings (#440)
* Fix generation of client uuid


* Remove all client properties when removed; Cleanup a few things


* Make all uuid function arguments use jack_uuid_t type; Fix warnings


* Rework uuids to never be int, more cleanup

+++ common/JackNetTool.cpp
@@ -884,7 +884,7 @@ namespace Jack
         for (int port_index = 0; port_index < fNPorts; port_index++) {
             if (fPortBuffer[port_index]) {
                 int res = opus_custom_decode_float(fOpusDecoder[port_index], fCompressedBuffer[port_index], fCompressedSizesByte[port_index], fPortBuffer[port_index], ((nframes == -1) ? fPeriodSize : nframes));
-                if (res < 0 || res != ((nframes == -1) ? fPeriodSize : nframes)) {
+                if (res < 0 || res != ((nframes == -1) ? (int)fPeriodSize : nframes)) {
                     jack_error("opus_custom_decode_float error fCompressedSizeByte = %d res = %d", fCompressedSizesByte[port_index], res);
                 }
             }
